fix(notification): fix language for notification

+++ app/src/main/java/com/tebet/mojual/data/models/request/MessageRequest.kt
@@ -1,7 +1,7 @@
 package com.tebet.mojual.data.models.request
 
 data class MessageRequest(
-    var profileId: Long? = null,
+    var profileId: Int? = null,
     var limit: Int? = null,
     var offset: Int? = null
 )
